requirements.txt
django==2.0.1
django-compressor==2.2
Pillow==5.0.0
whitenoise==3.3.1
sorl-thumbnail==12.4.1
psycopg2==2.7.3.2
dj-database-url==0.4.2

# Tests
factory_boy==2.9.2

# Docs
Sphinx
sphinx-autobuild
sphinxcontrib-django
sphinx_rtd_theme
